extract shuttle bus logic into its own service
